+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Add constructor to mock consumer ([370d5d6](https://github.com/tradewelltech/beavers/commit/370d5d68eb60662a110026ab7844fc3d9c6bf59b) by aandres).
 
+### Fixed
+
+- Fix offset resolution on end of topic ([ff76c35](https://github.com/tradewelltech/beavers/commit/ff76c3519d4ae36040cf138059952c9304bc1b3d) by aandres).
+
 ## [v0.7.0](https://github.com/tradewelltech/beavers/releases/tag/v0.7.0) - 2024-06-25
 
 <small>[Compare with v0.6.0](https://github.com/tradewelltech/beavers/compare/v0.6.0...v0.7.0)</small>